+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* CSP only allows scripts, styles and fonts from self which reflects GOV.UK production behaviour
 * Set the default CSP behaviour to be allow communication only to self
 * Remove webchat scripts from the CSP, these are now handled in [government-frontend](https://github.com/alphagov/government-frontend/pull/2643)
 * Remove `www.signin.service.gov.uk` from the CSP as it is no-longer used in GOV.UK